♻️refactor: 타입 변경
Long -> Integer.
방이 최대 3개 까지만 생길 것임으로.

+++ OneCoin/Server/src/main/java/OneCoin/Server/chat/chatMessage/entity/ChatMessage.java
@@ -17,7 +17,7 @@ public class ChatMessage {
     @JsonProperty("user_id")
     private Long userId;
     @JsonProperty("chat_room_id")
-    private Long chatRoomId;
+    private Integer chatRoomId;
     @JsonProperty("user_display_name")
     private String userDisplayName;
 }
